requirements.txt
kafka-python
pymongo
pandas
numpy
tensorflow
scikit-learn
flask
requests
beautifulsoup4
selenium
pyspark
